% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2025.Rmd.
\docType{data}
\name{authors2025}
\alias{authors2025}
\title{JSM 2025 author names}
\format{
A character vector of 5,515 JSM 2025 author names
}
\source{
This data was from the JSM 2025 program.
}
\usage{
authors2025
}
\description{
JSM 2025 author names
}
\keyword{datasets}