% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2025.Rmd.
\docType{package}
\name{jsm2025-package}
\alias{jsm2025}
\alias{jsm2025-package}
\title{A package to help you decide what talks to attend at JSM}
\description{
We wrangled the JSM program and decades of citation and coauthorship
data from Semantic Scholar and Arxiv.  We have also included functionality
for exporting your schedule as an ical that can be loaded into Google
Calendar or similar. The result is a package that streamlines the
process of finding talks that you may want to attend.
}
\seealso{
\code{\link{get_talks}}, \code{\link{get_coauthors}}, \code{\link{get_out_citations}}, \code{\link{get_in_citations}}, \code{\link{export_calendar_to_ics}}, \code{\link{export_calendar_to_csv}}
}
\author{
\strong{Maintainer}: Jacob Bien \email{jbien@usc.edu}

Other contributors:
\itemize{
  \item Yibin Xiong [contributor]
}

}
\keyword{internal}