% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-jsm2025.Rmd.
\docType{data}
\name{words}
\alias{words}
\title{JSM words used in titles}
\format{
A character vector of 4627 JSM words from titles
}
\source{
From the JSM 2025 program.
}
\usage{
words
}
\description{
JSM words used in titles
}
\keyword{datasets}